Move common backend functionality to new vantage6-backend-common package and clean up dependencies
Also added new package to pipelines
