+++ Makefile
@@ -47,6 +47,7 @@ uninstall:
 	pip uninstall -y vantage6-algorithm-tools
 	pip uninstall -y vantage6-common
 	pip uninstall -y vantage6-node
+	pip uninstall -y vantage6-backend-common
 	pip uninstall -y vantage6-server
 	pip uninstall -y vantage6-algorithm-store
 
@@ -56,6 +57,7 @@ install:
 	cd vantage6-algorithm-tools && pip install .
 	cd vantage6 && pip install .
 	cd vantage6-node && pip install .
+	cd vantage6-backend-common && pip install .
 	cd vantage6-server && pip install .
 	cd vantage6-algorithm-store && pip install .
 
@@ -65,6 +67,7 @@ install-dev:
 	cd vantage6-algorithm-tools && pip install -e .
 	cd vantage6 && pip install -e .[dev]
 	cd vantage6-node && pip install -e .[dev]
+	cd vantage6-backend-common && pip install -e .[dev]
 	cd vantage6-server && pip install -e .[dev]
 	cd vantage6-algorithm-store && pip install -e .[dev]
 
@@ -203,6 +206,10 @@ rebuild:
 	@echo "------------------------------------"
 	cd vantage6-node && make rebuild
 	@echo "------------------------------------"
+	@echo "         VANTAGE6 BACKEND COMMON    "
+	@echo "------------------------------------"
+	cd vantage6-backend-common && make rebuild
+	@echo "------------------------------------"
 	@echo "         VANTAGE6 SERVER            "
 	@echo "------------------------------------"
 	cd vantage6-server && make rebuild
@@ -217,6 +224,7 @@ publish:
 	cd vantage6-algorithm-tools && make publish
 	cd vantage6 && make publish
 	cd vantage6-node && make publish
+	cd vantage6-backend-common && make publish
 	cd vantage6-server && make publish
 	cd vantage6-algorithm-store && make publish
