+++ README.md
@@ -108,7 +108,9 @@ This repository is home to 6 PyPi packages:
 * [vantage6-algorithm-tools](https://pypi.org/project/vantage6-algorithm-tools) -> _Python tools to facilitate algorithm development_
 * [vantage6-node](https://pypi.org/project/vantage6-node) -> _Node application package_
 * [vantage6-server](https://pypi.org/project/vantage6-server) -> _Server application package_
-* [vantage6-common](https://pypi.org/project/vantage6-common) -> _Server application package_
+* [vantage6-algorithm-store](https://pypi.org/project/vantage6-algorithm-store) -> _Algorithm store application package_
+* [vantage6-common](https://pypi.org/project/vantage6-common) -> _Package with common vantage6 functions_
+* [vantage6-backend-common](https://pypi.org/project/vantage6-backend-common) -> _Package with functions common to central server and algorithm store_
 
 **Note that when using vantage6 you do not install the _server_ and _node_ packages. These are delivered to you in Docker images.**
