+++ vantage6-algorithm-store/requirements.txt
@@ -1,24 +1,12 @@
-# TODO figure out dependencies - here and in setup.py
-bcrypt==4.0.1
 flasgger==0.9.5
 flask==2.2.5
 Flask-Cors==3.0.10
-Flask-JWT-Extended==4.4.4
 Flask-Principal==0.4.0
-flask-mail==0.9.1
 flask-marshmallow==0.15.0
 Flask-RESTful==0.3.10
-Flask-SocketIO==5.3.6
 gevent==23.9.1
-ipython==8.10.0
-kombu==5.2.4
 marshmallow==3.19.0
-marshmallow-sqlalchemy==0.29.0
-PyJWT==2.6.0
-pyotp==2.8.0
-questionary==1.10.0
 requests==2.31.0
-requests-oauthlib==1.3.1
 schema==0.7.5
 SQLAlchemy==1.4.46
 werkzeug==3.0.1
