+++ vantage6-algorithm-store/setup.py
@@ -31,26 +31,15 @@ setup(
     packages=find_namespace_packages(),
     python_requires=">=3.10",
     install_requires=[
-        "bcrypt==4.0.1",
         "flasgger==0.9.5",
         "flask==2.2.5",
         "Flask-Cors==3.0.10",
-        "Flask-JWT-Extended==4.4.4",
-        "Flask-Mail==0.9.1",
         "Flask-Principal==0.4.0",
-        "Flask-RESTful==0.3.10",
         "flask-marshmallow==0.15.0",
-        "Flask-SocketIO==5.3.6",
+        "Flask-RESTful==0.3.10",
         "gevent==23.9.1",
-        "ipython==8.10.0",
-        "kombu==5.2.4",
         "marshmallow==3.19.0",
-        "marshmallow-sqlalchemy==0.29.0",
-        "PyJWT==2.6.0",
-        "pyotp==2.8.0",
-        "questionary==1.10.0",
         "requests==2.31.0",
-        "requests-oauthlib==1.3.1",
         "schema==0.7.5",
         "SQLAlchemy==1.4.46",
         "werkzeug==3.0.1",
@@ -59,13 +48,8 @@ setup(
     ],
     extras_require={"dev": ["coverage==6.4.4"]},
     package_data={
-        "vantage6.server": [
+        "vantage6.algorithm.store": [
             "__build__",
-            "_data/**/*.yaml",
-            "server_data/*.yaml",
         ],
-    },
-    entry_points={
-        "console_scripts": ["vserver-local=vantage6.server.cli.server:cli_server"]
     },
 )
