+++ vantage6-algorithm-store/vantage6/algorithm/store/resource/schema/output_schema.py
@@ -1,11 +1,11 @@
 from marshmallow import fields
 
+from vantage6.backend.common.resource.output_schema import BaseHATEOASModelSchema
 from vantage6.algorithm.store.model.algorithm import Algorithm
 from vantage6.algorithm.store.model.argument import Argument
 from vantage6.algorithm.store.model.database import Database
 from vantage6.algorithm.store.model.function import Function
 from vantage6.algorithm.store.model.vantage6_server import Vantage6Server
-from vantage6.server.resource.common.output_schema import BaseHATEOASModelSchema
 
 
 class HATEOASModelSchema(BaseHATEOASModelSchema):
