+++ vantage6-algorithm-store/vantage6/algorithm/store/resource/vantage6_server.py
@@ -16,7 +16,7 @@ from vantage6.algorithm.store.model.vantage6_server import (
 from vantage6.algorithm.store.resource import with_authentication
 
 # TODO move to common / refactor
-from vantage6.server.resource import AlgorithmStoreResources
+from vantage6.algorithm.store.resource import AlgorithmStoreResources
 
 
 module_name = logger_name(__name__)
