+++ vantage6-backend-common/vantage6/backend/common/globals.py
@@ -1,0 +1,3 @@
+# pagination settings
+DEFAULT_PAGE = 1
+DEFAULT_PAGE_SIZE = 10
