+++ vantage6-server/requirements.txt
@@ -12,10 +12,8 @@ gevent==23.9.1
 ipython==8.10.0
 kombu==5.2.4
 marshmallow==3.19.0
-marshmallow-sqlalchemy==0.29.0
 PyJWT==2.6.0
 pyotp==2.8.0
-questionary==1.10.0
 requests==2.31.0
 requests-oauthlib==1.3.1
 schema==0.7.5
