+++ vantage6-server/vantage6/server/globals.py
@@ -64,10 +64,6 @@ DEFAULT_EMAIL_FROM_ADDRESS = "noreply@vantage6.ai"
 # default time that token is valid in minutes
 DEFAULT_EMAILED_TOKEN_VALIDITY_MINUTES = 60
 
-# pagination settings
-DEFAULT_PAGE = 1
-DEFAULT_PAGE_SIZE = 10
-
 # default password policies
 DEFAULT_MAX_FAILED_ATTEMPTS = 5
 DEFAULT_INACTIVATION_MINUTES = 15
