+++ vantage6-server/vantage6/server/resource/algorithm_store.py
@@ -7,7 +7,7 @@ from http import HTTPStatus
 from sqlalchemy import or_
 
 from vantage6.server import db
-from vantage6.server.resource.common.pagination import Pagination
+from vantage6.backend.common.resource.pagination import Pagination
 from vantage6.server.resource.common.input_schema import AlgorithmStoreInputSchema
 from vantage6.server.permission import RuleCollection, Operation as P
 from vantage6.server.resource.common.output_schema import AlgorithmStoreSchema
