+++ vantage6-server/vantage6/server/resource/collaboration.py
@@ -5,7 +5,7 @@ from flask_restful import Api
 from http import HTTPStatus
 
 from vantage6.server import db
-from vantage6.server.resource.common.pagination import Pagination
+from vantage6.backend.common.resource.pagination import Pagination
 from vantage6.server.resource.common.input_schema import (
     CollaborationAddNodeSchema,
     CollaborationAddOrganizationSchema,
