+++ vantage6-server/vantage6/server/resource/task.py
@@ -26,7 +26,7 @@ from vantage6.server.resource.common.output_schema import (
     TaskWithRunAndResultSchema,
 )
 from vantage6.server.resource.common.input_schema import TaskInputSchema
-from vantage6.server.resource.common.pagination import Pagination
+from vantage6.backend.common.resource.pagination import Pagination
 from vantage6.server.resource.event import kill_task
