+++ vantage6-server/vantage6/server/resource/user.py
@@ -19,7 +19,7 @@ from vantage6.server.resource import (
     ServicesResources,
 )
 from vantage6.server.resource.common.input_schema import UserInputSchema
-from vantage6.server.resource.common.pagination import Pagination
+from vantage6.backend.common.resource.pagination import Pagination
 from vantage6.server.resource.common.output_schema import (
     UserSchema,
     UserWithPermissionDetailsSchema,
